Donor edits pickup location
https://trello.com/c/2gBYccFO

When a Donor's zipcode is supported, they'll need to edit their pickup
address.

This commit adds a `has_one :location` relationship to the User model.
There's nothing technical preventing Users from having more than one
location, but for the sake of simplicity we're limiting locations to a
1-to-1.

When a donor (supported or not) updates their address' zipcode, it will
update their Profiles "global" zipcode, which is used to determine
whether or not they're supported.