Supported Zipcode Registration Flow
https://trello.com/c/LP2n6hcq

This commit improves the registration flow.

First, potential donors enter their zipcode into the marketing page's
call to action.

This commit implements the flow for when a donor's zipcode is supported
by the program.

When potential donors are within the service area, they're prompted with
a registration form that asks for the typical `email` / `password`
combination, as well as their Name, Address, and Zipcode.

Once registration is complete, Users are redirected to their profile.

The bulk of the implementation depends on the `Registration` model,
which acts as a form object that wires together `User` and `Location`
instances.