Subscribe Unsupported Donors
https://trello.com/c/nhioUOIz

When a potential donor is in a not-yet-launched zipcode, allow them to
sign up to be notified once we launch in their area.

This commit introduces a `Subscription` model which represents a pairing
of emails and zipcodes.

Additionally, this commit removes failing test covered in another
branch.

A functioning implementation is [blocked].

[blocked]: https://trello.com/c/Rc1lNHcW/41-once-i-create-an-account-i-want-to-see-my-profile-so-i-know-what-information-i-have-provided-and-what-options-i-have-to-customiz#comment-56fea207e15c6b86877d710c