Render Header in `layouts/application`
Render a catch-all header partial from the Application layout.

This allows us to DRY-up our markup. If we'd like to split the `admin`
and `main` headers in the future, the `application/header` partial is a
great place to do that.

This commit removes the `.admin-nav` class, as it was a duplicate of the
now common `.main-nav`.