Enroll Donor's in Currently Scheduled Pickups
https://trello.com/c/nNxdjuZq

A `Donation` represents the relationship between a Donor with a location
and a `Zone`'s currently scheduled pickup.

When registering a new donor, they're automatically enrolled in their
zone's currently scheduled pickup, when applicable.

From the `/profile` page, donors are able to either confirm their
donation for the week or decline to donate.