Determine Donation status based on latest change
A donation can be both confirmed and declined.

Instead of blanking the opposing column whenever an action is taken, we
can preserve the timestamp of each action and determine current status
based on the latest action taken.

For example, if a donation is first confirmed, then declined, the
current status should reflect that it was declined at the timestamp the
action occurred, as well as maintain the timestamp that it was initally
confirmed at (for history's sake).

The reverse is also true.

This commit achieves this by overriding the `#*?` methods created by
`time_for_a_boolean` while still leveraging the rest of what that gem
has to offer.