Simplify Profile
https://trello.com/c/Oz6iPKen

This commit implements the design changes introduce in [5b4243c][].

It removes the `donations/current` partial in favor of a
`donations/donation` partial. Tests covering `donations/current` are
modified to provide coverage for `donations/donation`.

There are still unimplemented portions of the `/profile` page. They're
copied, unchanged, from the `high_voltage` version of `/profile` and
will be implemented in the future.

[5b4243c]: https://github.com/thoughtbot/freshfoodconnect/commit/5b4243c8da7be93865eeac3530d7280063813351