Extract the `donations/status` layout
This commit introduces the `donations/status` partial to be used as a
layout that yields a block.

The partial is responsible for showing the internationalized status for
the donation.

Since rendering partials as layouts has a verbose syntax, this commit
also introduces the `render_donation_status` helper to encapsulate the
magic incantation to correctly render a partial with a block.