Redirect cyclist to their latest assigned pickups
https://trello.com/c/wvpMTQHG

When cyclists volunteer, administrators will create account for them.

In the future, when there are more active zones, administrators will
assign cyclists to zones.

For the time being, any cyclist account that is created will be assigned
to the first created zone (which is currently `80205` in `staging` and
`production`).

This commit introduces the plumbing to assign a cyclist to a single
zone, maximizing future flexibility.