Cleanup `/users` styles and links
This commit cleans up the list of Users rendered at `/users`.

Prior to these changes, the `Cyclists` table had links on the left-most
column, while other tables had links and actions in the right-most
column. This commit moves the Cyclist link from the name to a `Register`
action in the right-most column.

When viewing the donors for a particular scheduled pickup, admins will
want to view a donor's donation history. This commit adds a link to
their `/donors/:id` page.