Attempt to resolve notification bug
Staging and Production are scheduled to execute `rake
confirmations:request` at the start of each day (00:00). This commit
introduces some leniency in the "48 hour" window so that a task kicked
off start-of-day Monday will send emails for a pickup scheduled at 08:00
Wednesday morning.

This commit replaces the Postgres-native `NOW()` call with a
Ruby-land `Time.current` value so that we can test and debug using
`Timecop`'s time freezing features.